+++ src/main/scala/chisel3/util/Counter.scala
@@ -3,6 +3,7 @@
 package chisel3.util
 
 import chisel3._
+import chisel3.internal.naming.chiselName  // can't use chisel3_ version because of compile order
 //import chisel3.core.ExplicitCompileOptions.Strict
 
 /** A counter module
@@ -10,6 +11,7 @@ import chisel3._
   * @param n number of counts before the counter resets (or one more than the
   * maximum output value of the counter), need not be a power of two
   */
+@chiselName
 class Counter(val n: Int) {
   require(n >= 0)
   val value = if (n > 1) Reg(init=0.U(log2Up(n).W)) else 0.U
@@ -53,6 +55,7 @@ object Counter
     * }
     * }}}
     */
+  @chiselName
   def apply(cond: Bool, n: Int): (UInt, Bool) = {
     val c = new Counter(n)
     var wrap: Bool = null
