+++ src/main/scala/chisel3/util/LFSR.scala
@@ -6,6 +6,7 @@
 package chisel3.util
 
 import chisel3._
+import chisel3.internal.naming.chiselName  // can't use chisel3_ version because of compile order
 //import chisel3.core.ExplicitCompileOptions.Strict
 
 // scalastyle:off magic.number
@@ -15,6 +16,7 @@ object LFSR16 {
     *
     * @param increment optional control to gate when the LFSR updates.
     */
+  @chiselName
   def apply(increment: Bool = true.B): UInt = {
     val width = 16
     val lfsr = Reg(init=1.U(width.W))
